requirements.txt
benfordslaw==1.0.2
matplotlib==3.4.2
pandas==1.3.2
numpy==1.20.3
streamlit==0.88.0
plotly==4.14.3
seaborn==0.11.2
Pillow==8.3.2
